package/commander: support Command interface for both root CLI and subcommands.
